+ Added different layers to the lobby scene for getting all options loaded into the game. * Because it's not finished, you can't actually start the game.

+++ Assets/Game Folder/Scripts/MatchStatistics.cs
@@ -265,7 +265,7 @@ public static class MatchStatistics
     /// Return an integer with the amount of minutes the match lasts.
     /// </summary>
     /// <returns></returns>
-    public static float GetMatchTimeInMinutes()
+    public static int GetMatchTimeInMinutes()
     {
         return matchTimeInMinutes;
     }
